# openapi.yml
openapi: 3.0.3
info:
  title: ACMS - Apartment Committee Management System
  description: |
    Multi-tenant SaaS platform for apartment/housing societies.
    Five role types: site_admin, apartment_admin, committee_head, committee_member, resident.
    Tenant isolation via `apartmentId` — all non-site-admin queries are scoped.
  version: 1.0.0
  contact:
    name: ACMS Team
servers:
  - url: /api
    description: API base path
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    InvoiceStatus:
      type: string
      enum: [paid, unpaid]
...
    SaaSInvoice:
      type: object
      properties:
        _id: { type: string }
        apartmentId: { type: string, description: Apartment ID or populated object }
        planId: { type: string, description: Plan ID or populated object }
        period: { type: string, example: '2026-07' }
        amount: { type: number }
        status: { $ref: '#/components/schemas/InvoiceStatus' }
        generatedAt: { type: string, format: date-time }
        dueDate: { type: string, format: date-time }
...
    Unit:
      type: object
      properties:
        _id: { type: string }
        apartmentId: { type: string }
        unitNumber: { type: string }
        residentUserId: { type: string, nullable: true, description: 'User ID or populated object' }
        status: { type: string, enum: [occupied, vacant] }
    CreateUnitRequest:
      type: object
      required: [unitNumber]
      properties:
        unitNumber: { type: string, example: A-101 }
        status: { type: string, enum: [occupied, vacant], default: vacant }
        residentUserId: { type: string, nullable: true }

    # ────────── Resident ──────────
    Resident:
      type: object
      properties:
        _id: { type: string }
        name: { type: string }
        identifier: { type: string }
        type: { type: string }
        unitId: { nullable: true }
        unitNumber: { type: string, nullable: true }
        status: { type: string }
        createdAt: { type: string, format: date-time }
    CreateResidentRequest:
      type: object
      required: [name, identifier, password]
      properties:
        name: { type: string }
        identifier: { type: string, description: Login ID (unit number) }
        password: { type: string, minLength: 6 }
        unitId: { type: string, nullable: true }
    UpdateResidentRequest:
      type: object
      properties:
        name: { type: string }
        status: { type: string, enum: [active, inactive] }
        unitId: { type: string, nullable: true }

    # ────────── Committee ──────────
    Committee:
      type: object
      properties:
        _id: { type: string }
        apartmentId: { type: string }
        name: { type: string }
        headUserId: { type: string, nullable: true, description: 'User ID or populated' }
        createdAt: { type: string, format: date-time }
        status: { type: string, enum: [active, inactive] }
    CreateCommitteeRequest:
      type: object
      required: [name]
      properties:
        name: { type: string, example: Maintenance Committee }
    CreateCommitteeHeadRequest:
      type: object
      required: [name, identifier, password]
      properties:
        name: { type: string }
        identifier: { type: string }
        password: { type: string, minLength: 6 }
...
paths:
...
  /admin/invoices:
    get:
      tags: [Site Admin]
      summary: List all SaaS invoices across all apartments
      security: [{ BearerAuth: [] }]
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SaaSInvoice'
...
  /apartment/invoices:
    get:
      tags: [Apartment Admin]
      summary: List SaaS invoices for the authenticated admin's apartment
      security: [{ BearerAuth: [] }]
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SaaSInvoice'

  # ══════════════════════════════════════════════════
  # APARTMENT ADMIN — Units & Residents
  # ══════════════════════════════════════════════════
  /apartment/units:
    get:
      tags: [Apartment Admin]
      summary: List all units in the apartment
      security: [{ BearerAuth: [] }]
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Unit'
    post:
      tags: [Apartment Admin]
      summary: Create a unit
      security: [{ BearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUnitRequest'
      responses:
        '201': { description: Unit created }
        '409': { description: Unit number already exists }
  /apartment/units/{id}:
    put:
      tags: [Apartment Admin]
      summary: Update a unit
      security: [{ BearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUnitRequest'
      responses:
        '200': { description: Unit updated }
        '404': { description: Unit not found }
    delete:
      tags: [Apartment Admin]
      summary: Delete a unit
      security: [{ BearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200': { description: Unit deleted }
  /apartment/residents:
    get:
      tags: [Apartment Admin]
      summary: List residents in the apartment
      security: [{ BearerAuth: [] }]
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Resident'
    post:
      tags: [Apartment Admin]
      summary: Create a resident account
      security: [{ BearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateResidentRequest'
      responses:
        '201': { description: Resident created }
        '409': { description: Identifier already exists }
  /apartment/residents/{id}:
    put:
      tags: [Apartment Admin]
      summary: Update a resident (name, status, unit assignment)
      security: [{ BearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateResidentRequest'
      responses:
        '200': { description: Resident updated }
        '404': { description: Resident not found }
    delete:
      tags: [Apartment Admin]
      summary: Delete a resident (frees unit)
      security: [{ BearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200': { description: Resident deleted }

  # ══════════════════════════════════════════════════
  # APARTMENT ADMIN — Committees
  # ══════════════════════════════════════════════════
  /apartment/committees:
    get:
      tags: [Apartment Admin]
      summary: List committees in the apartment
      security: [{ BearerAuth: [] }]
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Committee'
    post:
      tags: [Apartment Admin]
      summary: Create a committee
      security: [{ BearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCommitteeRequest'
      responses:
        '201': { description: Committee created }
        '409': { description: Committee name already exists }
  /apartment/committees/{id}:
    put:
      tags: [Apartment Admin]
      summary: Update committee name/status
      security: [{ BearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string }
                status: { type: string, enum: [active, inactive] }
      responses:
        '200': { description: Committee updated }
        '404': { description: Committee not found }
    delete:
      tags: [Apartment Admin]
      summary: Delete a committee
      security: [{ BearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200': { description: Committee deleted }
  /apartment/committees/{id}/head:
    post:
      tags: [Apartment Admin]
      summary: Create a Committee Head for a committee
      security: [{ BearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCommitteeHeadRequest'
      responses:
        '201': { description: Committee head created }
        '409': { description: Identifier already exists }

  # ══════════════════════════════════════════════════
  # COMMITTEE — General
  # ══════════════════════════════════════════════════
  /committees:
    get:
      tags: [Committee]
      summary: List committees (apartment_admin sees all, committee sees own)
      security: [{ BearerAuth: [] }]
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Committee'
  /committees/{id}:
    get:
      tags: [Committee]
      summary: Get a single committee by ID
      security: [{ BearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200': { description: Committee details }
        '404': { description: Committee not found }
...